make it possible to not have surface_de files and move functions
